Move out uniswap market price into separate hook
